+++ src/cli.ts
@@ -62,7 +62,10 @@ program
             "ts",
         ])
     )
-    .option("--pedantic", "require enum and union types to set tags")
+    .option(
+        "--pedantic",
+        "require enum and union types to set all tags in-order"
+    )
     .option(
         "--main <aliases...>",
         "space-separated list of types used to encode and decode messages"