+++ tests-corpus/valid-bare-schema/map-key-alias/out.gen.js
@@ -1,0 +1,32 @@
+import * as bare from "@bare-ts/lib"
+
+export function readKeyType(bc) {
+    return bare.readString(bc)
+}
+
+export function writeKeyType(bc, x) {
+    bare.writeString(bc, x)
+}
+
+export function readDict(bc) {
+    const len = bare.readUintSafe(bc)
+    const result = new Map()
+    for (let i = 0; i < len; i++) {
+        const offset = bc.offset
+        const key = readKeyType(bc)
+        if (result.has(key)) {
+            bc.offset = offset
+            throw new bare.BareError(offset, "duplicated key")
+        }
+        result.set(key, bare.readString(bc))
+    }
+    return result
+}
+
+export function writeDict(bc, x) {
+    bare.writeUintSafe(bc, x.size)
+    for(const kv of x) {
+        writeKeyType(bc, kv[0])
+        bare.writeString(bc, kv[1])
+    }
+}